implemented Window menu according Apple Guidelines
